[web] ♻️ refactor: remove playAudio function from useEffect dependencies in useVerificationButton
