+++ apps/web/src/ui/global/widgets/components/VerificationButton/StyledButton/index.tsx
@@ -1,28 +1,3 @@
-import { type ForwardedRef, forwardRef } from 'react'
-import { tv } from 'tailwind-variants'
+import { StyledButtonView } from './StyledButtonView'
 
-import type { ButtonProps } from '@/ui/global/widgets/components/Button/types'
-import { Button } from '@/ui/global/widgets/components/Button'
-
-const buttonStyles = tv({
-  base: 'w-64',
-  variants: {
-    color: {
-      green: 'border-gray-100 text-gray-900',
-      red: 'bg-red-700 text-gray-100',
-    },
-  },
-})
-
-export type StyledButtonProps = {
-  color: 'red' | 'green'
-} & ButtonProps
-
-const StyledButtonComponent = (
-  { color, ...buttonProps }: StyledButtonProps,
-  ref: ForwardedRef<HTMLButtonElement>,
-) => {
-  return <Button ref={ref} {...buttonProps} className={buttonStyles({ color })} />
-}
-
-export const StyledButton = forwardRef(StyledButtonComponent)
+export const StyledButton = StyledButtonView
