--- apps/web/src/ui/global/widgets/components/VerificationButton/VerificationButtonProps.ts
@@ -1,10 +1,0 @@
-import type { ClassNameValue } from 'tailwind-merge'
-
-export type VerificationButtonProps = {
-  onClick: () => void
-  isAnswered: boolean
-  isAnswerVerified: boolean
-  isAnswerCorrect: boolean
-  isLoading?: boolean
-  className?: ClassNameValue
-}
